Expand world-vis visualization
The aim is to expand the WorldVis visualization by allowing for any type
of world data.
We also aim to add functionality by increasing the number of
complementary visualizations. For
instance, adding the ability to compare the data for multiple selected
countries or by adding
new charts when a user clicks on a particular country. We are focusing
only on the visual
encoding and interaction idiom level of vis design since the domain and
data are based on the
input.